backend/requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
Flask==2.3.2
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
packaging==24.1
psycopg2-binary==2.9.7
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
